+++ pkg/combustion/templates/15-fips-setup.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -euo pipefail
+
+fips-mode-setup --enable